% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asCategories.R
\name{asCategories}
\alias{asCategories}
\title{Create a Categories measure wrapper.}
\usage{
asCategories(measure)
}
\arguments{
\item{measure.}{Measure to wrap. Must be either Ordinal or Nominal.}
}
\value{
Continuous variable wrapped around @param measure.
}
\description{
Method for constructing a Categories wrapper.
}
