% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asContinuous.R
\name{asContinuous}
\alias{asContinuous}
\title{Create a Continuous measure wrapper.}
\usage{
asContinuous(measure)
}
\arguments{
\item{measure.}{Measure to wrap. Must be either Numeric or Ordinal.}
}
\value{
Continuous variable wrapped around @param measure.
}
\description{
Method for constructing a Continuous wrapper.
}
