% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asCounts.R
\name{asCounts}
\alias{asCounts}
\title{Create a Counts measure wrapper.}
\usage{
asCounts(measure)
}
\arguments{
\item{measure.}{Measure to wrap. Must be either Numeric or Ordinal.}
}
\value{
Continuous variable wrapped around @param measure.
}
\description{
Method for constructing a Continuous wrapper.
}
