% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{assess}
\alias{assess}
\title{Assess a conceptual model to see if there is evidence for/in support of it in the data.}
\usage{
assess(conceptualModel, data)
}
\arguments{
\item{conceptualModel}{ConceptualModel to test.}

\item{data}{Dataset. Data to assess.}
}
\description{
Method for assessing if the data provides evidence for/against the conceptual model.
Fits and shows results of executing one or more statistical models for assessing the conceptual model.
Returns a script for running the statistical models
}
