% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_cardinality_from_data.R
\name{assign_cardinality_from_data}
\alias{assign_cardinality_from_data}
\title{Method to assigning cardinality from data}
\usage{
assign_cardinality_from_data(variable, data)
}
\arguments{
\item{variable}{AbstractVariable}

\item{data}{Dataset from which calculate cardinality for @param variable}
}
\description{
This function calculates cardinality from data and updates the variable's cardinality
}
\examples{
assign_cardinality_from_data()
}
