% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\docType{class}
\name{Associates-class}
\alias{Associates-class}
\alias{Associates}
\title{Associates class}
\description{
Class for Associates relationships.
Not called directly.
}
\section{Slots}{

\describe{
\item{\code{lhs}}{AbstractVariable. A variable that is associated with another.}

\item{\code{rhs}}{AbstractVariable. A variable that is associated with another.}
}}

\examples{
Associates()
}
