% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associates_with.R
\name{associates_with}
\alias{associates_with}
\title{Specify a Associates relationship}
\usage{
associates_with(lhs, rhs)
}
\arguments{
\item{lhs}{AbstractVariable.}

\item{rhs}{AbstractVariable.}
}
\description{
Method for constructing an associates with relationship between two variables.
Updates the relationships for each variable to include the newly constructed Associates relationship.
}
\examples{
associates_with()
}
