% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assume.R
\name{assume}
\alias{assume}
\title{Specify an Assumes statement}
\usage{
assume(relationship, conceptualModel)
}
\arguments{
\item{relationship}{Relationship to assume.}

\item{conceptualModel}{ConceptualModel to which we should add the Assumption.}
}
\description{
Method for constructing an assumption about a relationship.
Returns the Assumption created.
}
