% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\name{Assumption}
\alias{Assumption}
\title{Assumption class}
\usage{
Assumption(
  relationship = new("Relates", lhs = Unobserved(), rhs = Unobserved()),
  conceptualModel = ConceptualModel()
)
}
\description{
Class for Assumptions.
}
\section{Slots}{

\describe{
\item{\code{relationship.}}{Relationship to assume.}

\item{\code{conceptualModel}}{ConceptualModel to which this Assumption belongs.}
}}

