% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\docType{class}
\name{AtMost-class}
\alias{AtMost-class}
\alias{AtMost}
\title{AtMost class}
\description{
Wrapper class for upperbound of number values. Extends NumberValue class.
}
\section{Slots}{

\describe{
\item{\code{value}}{Number value}
}}

\examples{
AtMost(3), could be 0, 1, 2, xor 3 value!
}
