% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\name{AtMost}
\alias{AtMost}
\title{AtMost class}
\usage{
AtMost(value = as.integer(1))
}
\description{
Wrapper class for upperbound of number values. Extends NumberValue class.
}
\section{Slots}{

\describe{
\item{\code{value}}{Number value}
}}

\examples{
AtMost(as.integer(1))
}
