% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cardinality.R
\name{calculate_cardinality_from_data}
\alias{calculate_cardinality_from_data}
\title{Method to calculate cardinality from data}
\usage{
calculate_cardinality_from_data(variable, data)
}
\arguments{
\item{variable}{AbstractVariable}

\item{data}{Dataset from which calculate cardinality for @param variable}
}
\description{
Method for calculating cardinality
}
