% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\docType{class}
\name{Categories-class}
\alias{Categories-class}
\alias{Categories}
\title{Categories class}
\description{
Class for (ordered, unordered) Categories
Inherits from Measure.
}
\section{Slots}{

\describe{
\item{\code{name}}{character. Name of measure, corresponds to column name in data.}
}}

