% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories.R
\name{categories}
\alias{categories}
\title{Create a Categories measure}
\usage{
categories(unit, name, cardinality, order, numberOfInstances = 1)
}
\arguments{
\item{unit}{Unit object. Unit that has/contributes the categories measure.}

\item{name}{Character. Name of measure, corresponds to column name in data.}

\item{numberOfInstances}{Integer or AbstractVariable or AtMost or Per. Number of instances of the measure the @param unit has. Default is 1.}

\item{cardinality.}{Integer. Optional. Only required if no data is assigned. Number of unique categories. If @param order is provided, @param cardinality is not needed and will be set to the length of @order.}

\item{order.}{List. Optional. List of categories in order from "lowest" to "highest".}
}
\value{
handle to Categories measure
}
\description{
Method for constructing a categories measure through a Unit.
}
