% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\docType{class}
\name{Causes-class}
\alias{Causes-class}
\alias{Causes}
\title{Causes class}
\description{
Class for Causes relationships.
Not called directly.
}
\section{Slots}{

\describe{
\item{\code{cause}}{AbstractVariable. Variable that causes another.}

\item{\code{effect}}{AbstractVariable. Variable that is caused by @slot cause.}
}}

