% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causes.R
\name{causes}
\alias{causes}
\title{Specify a causes relationship}
\usage{
causes(cause, effect)
}
\arguments{
\item{cause}{AbstractVariable. The cause.}

\item{effect}{AbstractVariable. The effect.}
}
\description{
Method for constructing a causes relationship between two variables.
Updates the relationships for each variable to include the newly constructed Causes relationship.
}
\examples{
causes()
}
