% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_conceptual_relationships.R
\name{check_conceptual_relationships}
\alias{check_conceptual_relationships}
\title{Construct graph of all relationships}
\usage{
check_conceptual_relationships(conceptual_gr)
}
\arguments{
\item{all_relationships}{List of all relationships (edges) to include in this graph}

\item{vars}{List of AbstractVariables involved in the graph}
}
\value{
Tuple of graphs: (conceptual graph, data measurement graph)
}
\description{
This function constructs a graph representing all relationships
}
\examples{
check_conceptual_relationships()
}
