% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAndRefine.R
\name{checkAndRefineConceptualModel}
\alias{checkAndRefineConceptualModel}
\title{Check and refine conceptual model through interactive disambiguation}
\usage{
checkAndRefineConceptualModel(conceptualModel, iv, dv)
}
\description{
Check and refine conceptual model through interactive disambiguation
}
