% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkConceptualModel.R
\name{checkConceptualModel}
\alias{checkConceptualModel}
\title{Check Conceptual Model (graph) is valid, meaning there are no cycles and that the @param dv does not cause @param iv
Use once have a query (i.e., iv and dv) of interest}
\usage{
checkConceptualModel(conceptualModel, iv, dv)
}
\arguments{
\item{conceptualModel}{ConceptualModel. Causal graph to validate.}

\item{iv}{AbstractVariable. Independent variable that should cause (directly or indirectly) the @param dv.}

\item{dv}{AbstractVariable. Dependent variable that should not cause the @param iv.}
}
\value{
TRUE if the Conceptual Model is valid, FALSE otherwise
}
\description{
This function validates the Conceptual Model
}
