% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkEquals.R
\name{checkEquals}
\alias{checkEquals}
\title{Checks if two values, treated as integers, are equal.}
\usage{
checkEquals(variable, value)
}
\description{
Returns TRUE if the values are equal. FALSE otherwise.
}
