% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkGreaterThan.R
\name{checkGreaterThan}
\alias{checkGreaterThan}
\title{Checks if a value is greater than another value.}
\usage{
checkGreaterThan(lhs, rhs)
}
\description{
Returns TRUE if @param lhs is greater than @param rhs.
}
