% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\docType{class}
\name{Compares-class}
\alias{Compares-class}
\alias{Compares}
\title{Compares class}
\description{
Class for comparison relationships.
Not called directly.
}
\section{Slots}{

\describe{
\item{\code{variable}}{AbstractVariable. Variable that is being compared.}

\item{\code{condition}}{character. Condition to filter values of @slot variable on.}
}}

