% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\name{ConceptualModel}
\alias{ConceptualModel}
\title{ConceptualModel class}
\usage{
ConceptualModel(
  variables = list(),
  relationships = list(),
  graph = dagitty({
     ""
 })
)
}
\description{
Class for Conceptual Models
}
