% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition.R
\name{condition}
\alias{condition}
\title{Create a condition}
\usage{
condition(unit, name, cardinality, order, numberOfInstances)
}
\arguments{
\item{unit}{Unit object. The condition was assigned to a Unit instance.}

\item{name}{Character. Name of measure, corresponds to column name in data.}

\item{numberOfInstances}{Integer or AbstractVariable or AtMost or Per. Number of instances of the measure the @param unit has. Default is 1.}

\item{cardinality.}{Integer. Optional. Only required if no data is assigned.}
}
\value{
Has Relationship representing Unit having the Nominal Measure.
}
\description{
Method for constructing a condition through a Unit
}
