% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\docType{class}
\name{Continuous-class}
\alias{Continuous-class}
\alias{Continuous}
\title{Continuous class}
\description{
Class for Continuous measures (e.g., scores, temperature, time)
Inherits from Measure.
}
\section{Slots}{

\describe{
\item{\code{name}}{character. Name of measure, corresponds to column name in data.}
}}

