% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous.R
\name{continuous}
\alias{continuous}
\title{Create a Continuous measure}
\usage{
continuous(unit, name, numberOfInstances = 1)
}
\arguments{
\item{unit}{Unit object. Unit that has/contributes the Continuous measure.}

\item{name}{Character. Name of measure, corresponds to column name in data.}
}
\value{
handle to Continuous measure.
}
\description{
Method for constructing a Continuous measure through a Unit.
}
