% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\docType{class}
\name{Counts-class}
\alias{Counts-class}
\alias{Counts}
\title{Counts class}
\description{
Wrapper class for Counts.
Not called directly. All measures are declared through Units as either Numeric, Ordinal, or Nominal types.
}
\section{Slots}{

\describe{
\item{\code{measure}}{Measure to wrap.}
}}

