% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{counts}
\alias{counts}
\title{Create a Counts measure}
\usage{
counts(unit, name, numberOfInstances = 1)
}
\arguments{
\item{unit}{Unit object. Unit that has/contributes the Counts measure.}

\item{name}{Character. Name of measure, corresponds to column name in data.}
}
\value{
handle to Counts measure.
}
\description{
Method for constructing a Counts measure through a Unit.
}
