% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\docType{class}
\name{Dataset-class}
\alias{Dataset-class}
\alias{Dataset}
\title{Dataset class}
\description{
Class for Dataset
}
\section{Slots}{

\describe{
\item{\code{df}}{DataFrame.}
}}

\examples{
Dataset()
}
