% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decreases.R
\name{decreases}
\alias{decreases}
\title{Specify a decreasing Compares relationship.}
\usage{
decreases(variable)
}
\arguments{
\item{variable}{numericORordinal. The variable decreasing.}
}
\description{
Returns the Compares relationship constructed.
}
