% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\name{Design}
\alias{Design}
\title{Design class}
\usage{
Design(relationships = list(), ivs = NULL, dv = NULL, source = NULL)
}
\description{
Class for Study designs.
}
\section{Slots}{

\describe{
\item{\code{relationships}}{List of relationships between variables.}

\item{\code{ivs}}{List of AbstractVariables. Varibale that are independent variables.}

\item{\code{dv}}{AbstractVariable. Variable that is the dependent variable.}

\item{\code{source}}{Data frame containing data for Design.}
}}

