% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambiguateStatisticalModel.R
\name{disambiguateStatisticallModel}
\alias{disambiguateStatisticallModel}
\title{Disambiguate statistical model candidates}
\usage{
disambiguateStatisticallModel(inputFilePath)
}
\description{
Disambiguate statistical model candidates
}
