% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equals.R
\name{equals}
\alias{equals}
\title{Specify an equals Compares relationship.}
\usage{
equals(variable, value)
}
\arguments{
\item{variable}{nominalORordinal The variable whose value may equal @param value.}

\item{value}{integerORnumericORcharacter. The value to look for/filter on in @param variable.}
}
\description{
Returns the Compares relationship constructed.
}
