% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\docType{class}
\name{Exactly-class}
\alias{Exactly-class}
\alias{Exactly}
\title{Exactly class}
\description{
Wrapper class for exact number values. Extends NumberValue class.
}
\section{Slots}{

\describe{
\item{\code{value}}{Number value}
}}

\examples{
Exactly(3)
}
