% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\name{Exactly}
\alias{Exactly}
\title{Exactly class}
\usage{
Exactly(value = as.integer(1))
}
\description{
Wrapper class for exact number values. Extends NumberValue class.
}
\section{Slots}{

\describe{
\item{\code{value}}{Number value}
}}

\examples{
Exactly(as.integer(3))
}
