% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processQuery.R
\name{generateConceptualModelJSON}
\alias{generateConceptualModelJSON}
\title{Output info to JSON}
\usage{
generateConceptualModelJSON(conceptualModel, path = "input.json")
}
\arguments{
\item{path}{should be different from "input.json" only for development/testing}
}
\description{
Writes info to a JSON file
}
