% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processQuery.R
\name{generateDisambiguationChoices}
\alias{generateDisambiguationChoices}
\title{Returns list of disambiguation choices}
\usage{
generateDisambiguationChoices(relationship, tag)
}
\arguments{
\item{relationship}{is a Relates or Causes object}

\item{tag}{is "Assume" or "Hypothesize"}
}
\description{
Returns list of disambiguation choices
}
