% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processQuery.R
\name{generateConceptualModelJSON}
\alias{generateConceptualModelJSON}
\title{Output info to JSON}
\usage{
generateConceptualModelJSON(conceptualModel, dv, path)
}
\arguments{
\item{dv}{AbstractVariable to write to JSON.}
}
\description{
Writes info to a JSON file
}
