% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStatisticalModels.R
\name{generateStatisticalModelJSON}
\alias{generateStatisticalModelJSON}
\title{Output info to JSON}
\usage{
generateStatisticalModelJSON(
  confounders,
  interactions,
  randomEffects,
  familyLinkFunctions,
  iv,
  dv,
  path
)
}
\arguments{
\item{confounders}{list of confounders to include in the statistical model.}

\item{interactions}{list of interaction effects to consider including in the statistical model. Optional.}

\item{randomEffects}{list of random effects to include in the statistical model to maximize generalizability. Optional.}

\item{familyLinkFunctions}{list of family and link functions to consider.}

\item{iv}{AbstractVariable whose influence on @param dv we are interested in.}

\item{dv}{Continuous, Counts, or Categories.}

\item{path}{Path-like or character. Path to write out the JSON.}
}
\description{
Writes info to a JSON file
}
