% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_vars.R
\name{get_all_vars}
\alias{get_all_vars}
\title{Get all the variables in the design}
\usage{
get_all_vars(design)
}
\arguments{
\item{design}{Design.}
}
\value{
List of all variables
}
\description{
This function gathers a list of all variables in a study design.
}
\examples{
infer_has_relationships()
}
