% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cardinality.R
\name{get_cardinality}
\alias{get_cardinality}
\title{Method to get cardinality}
\usage{
get_cardinality(variable)
}
\arguments{
\item{variable}{AbstractVariable to inspect}
}
\description{
Method for getting the cardinality of an AbstractVariable
}
\examples{
get_cardinality()
}
