% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferConfounders.R
\name{getMediators}
\alias{getMediators}
\title{Find mediators.}
\usage{
getMediators(conceptualModel, x, y)
}
\arguments{
\item{x}{character. Name of AbstractVariable in @param graph.}

\item{y}{character. Name of AbstractVariable in @param graph.}
}
\value{
list of Observed mediators.
}
\description{
This function finds Observed mediators on paths between @param x and @param y
Helper function to follow recommendations from Cinelli, Forney, and Pearl 2020
}
