% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processQuery.R
\name{getVariable}
\alias{getVariable}
\title{Returns handle to variable whose name we know}
\usage{
getVariable(conceptualModel, name)
}
\arguments{
\item{conceptualModel}{ConceptualModel with a variable of interest.}

\item{name}{character. Name of variable we would like a handle to.}
}
\description{
Helper function for navigating and updating a Conceptual Model. Returns handle to a variable or NULL otherwise.
precondition: There must be a variable with the @param name in the @param conceptualModel.
}
