% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\docType{class}
\name{Has-class}
\alias{Has-class}
\alias{Has}
\title{Has class}
\description{
Class for Has relationships between AbstractVariables.
Not called directly. For internal purposes only.
}
\section{Slots}{

\describe{
\item{\code{variable}}{AbstractVariable. Variable that has another Variable. For example, the Unit that has a Measure.}

\item{\code{measure}}{AbstractVariable. Variable that @slot variable has. For example, the Measure a Unit has.}

\item{\code{repetitions}}{NumberValue. Number of times that @slot variable has @slot measure.}

\item{\code{according_to}}{AbstractVariable. Variable whose unique values differentiate the repeated instances of @slot measure.}
}}

\examples{
Has()
}
