% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has.R
\name{has}
\alias{has}
\title{Create a has relationship}
\usage{
has(unit, measure, number_of_instances = 1)
}
\arguments{
\item{unit}{Unit. The Unit that has @param measure.}

\item{measure}{Measure. The Measure that @param unit has.}

\item{number_of_instances}{Integer or AbstractVariable or AtMost or Per. Number of instances of the @param measure the @param unit has. Default is 1.}
}
\description{
Method for specifying and creating a has relationship between Unit and Measure/
Not called directly. For internal purposes only.
}
\examples{
has()
}
