% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\name{Hypothesis}
\alias{Hypothesis}
\title{Hypothesis class}
\usage{
Hypothesis(relationship, conceptualModel)
}
\description{
Class for Hypotheses.
}
\section{Slots}{

\describe{
\item{\code{relationship.}}{Relationship to hypothesize.}

\item{\code{conceptualModel}}{ConceptualModel to which this Hypothesis belongs.}
}}

