% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/increases.R
\name{increases}
\alias{increases}
\title{Specify an increasing Compares relationship.}
\usage{
increases(variable)
}
\arguments{
\item{variable}{numericORordinal. The variable increasing.}
}
\description{
Returns the Compares relationship constructed.
}
