% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_has_relationships.R
\name{infer_has_relationships}
\alias{infer_has_relationships}
\title{Infer has relationships from a study design}
\usage{
infer_has_relationships(design)
}
\arguments{
\item{design}{}
}
\description{
This function infers has relationships from a study design.
}
\examples{
infer_has_relationships()
}
