% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infers_main_effects_with_explanations.R
\name{infer_main_effects_with_explanations}
\alias{infer_main_effects_with_explanations}
\title{Infers main effects.}
\usage{
infer_main_effects_with_explanations(causal_gr, associative_gr, design)
}
\arguments{
\item{causal_gr}{Graph representing causal relationships between variables.}

\item{associative_gr}{Graph representing associative relationships between variables.}

\item{design}{Design}
}
\value{
List of main effects and their explanations
}
\description{
This function infers main effects from a conceptual graph.
}
