% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_model.R
\name{infer_model}
\alias{infer_model}
\title{Infer Model Function}
\usage{
infer_model(design)
}
\arguments{
\item{design}{}
}
\description{
This function infers a statistical model from a study design. Calls infer_statistical_model_from_design
}
\keyword{model}
\keyword{statistical}
