% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_random_effects_with_explanations.R
\name{infer_random_effects_with_explanations}
\alias{infer_random_effects_with_explanations}
\title{Infers random effects.}
\usage{
infer_random_effects_with_explanations(
  measurement_gr,
  nests_gr,
  design,
  all_relationships,
  main_effects,
  interaction_effects
)
}
\arguments{
\item{measurement_gr}{Graph representing data measurement relationships.}

\item{nests_gr}{Graph representing nesting relationships between variables.}

\item{design}{Design}

\item{main_effects}{List of candidate main effects}

\item{interaction_effects}{List of candidate interaction effects}
}
\value{
List of random effects and their explanations
}
\description{
This function infers random effects from data measurement relationships.
}
