% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_statistical_model_from_design.R
\name{infer_statistical_model_from_design}
\alias{infer_statistical_model_from_design}
\title{Infer Statistical Model from Design Function}
\usage{
infer_statistical_model_from_design(design)
}
\arguments{
\item{design}{}
}
\description{
This function infers a statistical model from a study design.
}
\examples{
infer_statistical_model_from_design()
}
\keyword{model}
\keyword{statistical}
