% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferConfounders.R
\name{inferConfounders}
\alias{inferConfounders}
\title{Infer confounders to account for when assessing the direct effect of @param iv on @param dv.}
\usage{
inferConfounders(conceptualModel, iv, dv)
}
\arguments{
\item{conceptualModel}{ConceptualModel. With causal graph to use in order to identify confounders.}

\item{iv}{AbstractVariable. Independent variable that should cause (directly or indirectly) the @param dv.}

\item{dv}{AbstractVariable. Dependent variable that should not cause the @param iv.}
}
\value{
list of confounders to include.
}
\description{
This function findings confounders to account for.
Follows recommendations from Cinelli, Forney, and Pearl 2020
}
