% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferFamilyLinkFunctions.R
\name{inferFamilyFunctions}
\alias{inferFamilyFunctions}
\title{Find family functions.}
\usage{
inferFamilyFunctions(dv)
}
\arguments{
\item{dv}{AbstractVariable. Variable whose outcome we want to assess.}
}
\value{
list of possible family functions
}
\description{
This function finds possible family functions.
Helper function to infer family and link function pairs
}
