% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferFamilyLinkFunctions.R
\name{inferFamilyLinkFunctions}
\alias{inferFamilyLinkFunctions}
\title{Infer possible family and link functions for statistical model.}
\usage{
inferFamilyLinkFunctions(dv)
}
\arguments{
\item{dv}{ContinuousORCountsORCategories. Wrapped Measure whose outcome we want to assess.}
}
\description{
Infers possible family and link function pairs for dependent variable
Returns a list of possible family and link function pairs.
}
