% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferFamilyLinkFunctions.R
\name{inferLinkFunctions}
\alias{inferLinkFunctions}
\title{Find link function for a given family function.}
\usage{
inferLinkFunctions(family)
}
\arguments{
\item{family}{character. Name of family function for which we want to infer possible link functions.}
}
\value{
list of possible link functions
}
\description{
This function finds possible link functions for a family function.
Helper function to infer family and link function pairs
}
