% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferRandomEffects.R
\name{inferRandomEffects}
\alias{inferRandomEffects}
\title{Infers random effects.}
\usage{
inferRandomEffects(conceptualModel, iv, dv)
}
\arguments{
\item{measurement_gr}{Graph representing data measurement relationships.}

\item{nests_gr}{Graph representing nesting relationships between variables.}

\item{design}{Design}

\item{main_effects}{List of candidate main effects}

\item{interaction_effects}{List of candidate interaction effects}
}
\value{
List of random effects and their explanations
}
\description{
This function infers random effects from data measurement relationships.
}
