% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\docType{class}
\name{Interacts-class}
\alias{Interacts-class}
\alias{Interacts}
\title{Interacts"class}
\description{
Class for representing Interacts"effects.
Not called directly. All interactions are declared through interacts.
}
\section{Slots}{

\describe{
\item{\code{name}}{Character name (shorthand) for the interaction effect}

\item{\code{units}}{List of Units, may only be one unit if the moderation is constructed from Measures from the same Unit}

\item{\code{variables}}{List of AbstractVariables or UnobservedVariables that interact with each other}
}}

