% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interacts.R
\name{interacts}
\alias{interacts}
\title{Specify an interaction relationship}
\usage{
interacts(conceptualModel, ..., dv)
}
\arguments{
\item{conceptualModel}{ConceptualModel.}

\item{...}{AbstractVariables.}

\item{dv}{AbstractVariable.}
}
\description{
Method for constructing an interaction between two or more variables.
Returns the Interacts"constructed.
}
