% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moderates.R
\name{moderates,AbstractVariable,list,AbstractVariable-method}
\alias{moderates,AbstractVariable,list,AbstractVariable-method}
\title{Specify a Moderates relationship}
\usage{
\S4method{moderates}{AbstractVariable,list,AbstractVariable}(var, moderator, on)
}
\arguments{
\item{var}{AbstractVariable. Variable that moderates the effect of @param moderator on @param on variable.}

\item{moderator}{List of AbstractVaraibles.}

\item{on}{AbstractVariable.}
}
\description{
Method for constructing a moderates relationship between two variables.
Updates the relationships for each variable to include the newly constructed Moderates relationship.
}
