% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moderates.R
\docType{class}
\name{Moderates-class}
\alias{Moderates-class}
\title{Moderates class}
\arguments{
\item{moderator}{List of AbstractVariables. Variables that moderate each other's effect on the @param on variable.}

\item{on}{AbstractVariable. The outcome variable.}
}
\description{
Class for Moderates relationships.
Not called directly.
}
\examples{
Moderates()
}
