% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moderates.R
\name{moderates}
\alias{moderates}
\title{Specify a Moderates relationship}
\usage{
moderates(var, moderator, on)
}
\arguments{
\item{var}{AbstractVariable. Variable that moderates the effect of @param moderator on @param on variable.}

\item{moderator}{AbstractVariable.}

\item{on}{AbstractVariable.}
}
\description{
Method for constructing a moderates relationship between two variables.
Updates the relationships for each variable to include the newly constructed Moderates relationship.
}
\examples{
moderates()
}
