% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\docType{class}
\name{ModerationNominal-class}
\alias{ModerationNominal-class}
\alias{ModerationNominal}
\title{Moderation Nominal class}
\description{
Class for representing Moderation (or Moderation) effects as nominal variables.
Not called directly. All moderations are declared through moderates.
}
\section{Slots}{

\describe{
\item{\code{units}}{List of Units, may only be one unit if the moderation is constructed from Measures from the same Unit}

\item{\code{name}}{Character name (shorthand) for the interaction/moderation effect}

\item{\code{cardinality}}{Integer for cardinality.}

\item{\code{moderators}}{List of AbstractVariables that moderate each other}

\item{\code{categories}}{List of categories that are the product of interacting the moderators.}
}}

