% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\docType{class}
\name{Nests-class}
\alias{Nests-class}
\alias{Nests}
\title{Nests class}
\description{
Class for Nesting relationships.
Not called directly.
}
\section{Slots}{

\describe{
\item{\code{base}}{AbstractVariable. Variable that is nested within another.}

\item{\code{group}}{AbstractVariable. Variable that contains multiple instances of @slot base.}
}}

