% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nests.R
\name{nests}
\alias{nests}
\title{Specify a Nests relationship}
\usage{
nests(base, group)
}
\arguments{
\item{base}{Unit. Variable that is nested within another.}

\item{group}{Unit. Variable that contains multiple instances of @param base.}
}
\description{
Method for constructing a nesting relationship between two variables.
Updates the relationships for each variable to include the newly constructed Nests relationship.
}
