% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\docType{class}
\name{Nominal-class}
\alias{Nominal-class}
\alias{Nominal}
\title{Nominal class}
\description{
Class for Nominal measures, inherits from Measure.
Not called directly. All measures are declared through Units.
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of measure, corresponds to column name if assigning data.}

\item{\code{cardinality}}{Integer for cardinality.}

\item{\code{categories}}{List of categories.}

\item{\code{isInteraction}}{Logical. True if variable is an interaction. False otherwise.}
}}

\examples{
Nominal()
}
