% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nominal.R
\name{nominal}
\alias{nominal}
\title{Create a nominal measure}
\usage{
nominal(unit, name, cardinality, number_of_instances)
}
\arguments{
\item{unit}{Unit object. Unit that has/contributes the nominal measure.}

\item{name}{Character. Name of measure, corresponds to column name in data.}

\item{number_of_instances}{Integer or AbstractVariable or AtMost or Per. Number of instances of the measure the @param unit has. Default is 1.}

\item{cardinality.}{Integer. Optional. Only required if no data is assigned.}
}
\value{
Has Relationship representing Unit having the Nominal Measure.
}
\description{
Method for constructing a nominal measure through a Unit.
}
\examples{
nominal()
}
