% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notEquals.R
\name{notEquals}
\alias{notEquals}
\title{Specify a not notEquals Compares relationship.}
\usage{
notEquals(variable, value)
}
\arguments{
\item{variable}{nominalORordinal The variable whose value may not equal @param value.}

\item{value}{integerORnumericORcharacter. The value to look for/filter on in @param variable.}
}
\description{
Returns the Compares relationship constructed.
}
