% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\docType{class}
\name{NumberValue-class}
\alias{NumberValue-class}
\alias{NumberValue}
\title{Number Value class}
\description{
Abstract super class for number values, used for number_of_instances that a Unit has of a Measure
}
\section{Slots}{

\describe{
\item{\code{value}}{Number value}
}}

\examples{
NumberValue()
}
