% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\docType{class}
\name{Numeric-class}
\alias{Numeric-class}
\alias{Numeric}
\title{Numeric class}
\description{
Class for Numeric measures, inherits from Measure.
Not called directly. All measures are declared through Units.
}
\section{Slots}{

\describe{
\item{\code{name}}{Character. Name of measure, corresponds to column name in data.}
}}

\examples{
Numeric()
}
