% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric.R
\name{numeric}
\alias{numeric}
\title{Create a numeric measure}
\usage{
numeric(unit, name, number_of_instances = 1)
}
\arguments{
\item{unit}{Unit object. Unit that has/contributes the numeric measure.}

\item{name}{Character. Name of measure, corresponds to column name in data.}

\item{number_of_instances}{Integer or AbstractVariable or AtMost or Per. Number of instances of the measure the @param unit has. Default is 1.}
}
\value{
Has Relationship representing Unit having the Nominal Measure.
}
\description{
Method for constructing a numeric measure through a Unit.
}
\examples{
numeric()
}
