% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\docType{class}
\name{OrderedCategories-class}
\alias{OrderedCategories-class}
\alias{OrderedCategories}
\title{Ordered Categories class}
\description{
Class for ordered Categories
Inherits from Measure.
}
\section{Slots}{

\describe{
\item{\code{name}}{character. Name of measure, corresponds to column name in data.}

\item{\code{cardinality}}{integer. Number of unique categories. If @param order is provided, @param cardinality is not needed and will be set to the length of @param order}

\item{\code{order}}{list. Optional. List of categories in order from "lowest" to "highest"}
}}

