% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\docType{class}
\name{Ordinal-class}
\alias{Ordinal-class}
\alias{Ordinal}
\title{Ordinal class}
\description{
Class for Ordinal measures, inherits from Measure.
Not called directly. All measures are declared through Units.
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of measure, corresponds to column name in data.}

\item{\code{cardinality}}{Integer for cardinality.}

\item{\code{order}}{Ordered list of categories.}
}}

