% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal.R
\name{ordinal}
\alias{ordinal}
\title{Create an ordinal measure}
\usage{
ordinal(unit, name, order, number_of_instances)
}
\arguments{
\item{unit}{Unit object. Unit that has/contributes the ordinal measure.}

\item{name}{Character. Name of measure, corresponds to column name in data.}

\item{order}{List. Ordered list of categories.}

\item{number_of_instances}{Integer or AbstractVariable or AtMost or Per. Number of instances of the measure the @param unit has. Default is 1.}

\item{cardinality.}{Integer. Optional. Only required if no data is assigned. If provided, checked to make sure @param cardinality == length(@param order)}
}
\value{
Has Relationship representing Unit having the Nominal Measure.
}
\description{
Method for constructing an ordinal measure through a Unit.
}
\examples{
ordinal()
}
