% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\name{Participant}
\alias{Participant}
\title{Participant class}
\usage{
Participant(name, cardinality = as.integer(0))
}
\description{
Class for Participant variables
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of Participant, corresponds to column name if assigning data}

\item{\code{integer}}{Integer for cardinality, optional. Only required if no data is assigned}
}}

