% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\docType{class}
\name{Per-class}
\alias{Per-class}
\alias{Per}
\title{Per class}
\description{
Wrapper class for ratios between number values. Extends NumberValue class.
}
\section{Slots}{

\describe{
\item{\code{number}}{NumberValue}

\item{\code{variable}}{AbstractVariable that contains}
}}

