% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processQuery.R
\name{processQuery}
\alias{processQuery}
\title{Elicit any additional information about Conceptual Model (graph) if necessary}
\usage{
processQuery(conceptualModel, iv, dv, data)
}
\arguments{
\item{conceptualModel}{ConceptualModel. Contains causal graph to process.}

\item{iv}{AbstractVariable. Independent variable whose influence on @param dv we want to assess.}

\item{dv}{AbstractVariable. Dependent variable.}
}
\description{
This function disambiguates the Conceptual Model before using it in future query steps.
}
