% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriveStatisticalModel.R
\name{deriveStatisticalModel}
\alias{deriveStatisticalModel}
\title{Elicit additional information about to narrow the space of candidate statistical models to a final output one.}
\usage{
deriveStatisticalModel(
  confounders,
  interactions,
  randomEffects,
  familyLinkFunctions,
  iv,
  dv,
  data
)
}
\arguments{
\item{confounders}{list of confounders to include in the statistical model.}

\item{interactions}{list of interaction effects to consider including in the statistical model. Optional.}

\item{randomEffects}{list of random effects to include in the statistical model to maximize generalizability. Optional.}

\item{familyLinkFunctions}{list of family and link functions to consider.}

\item{iv}{AbstractVariable whose influence on @param dv we are interested in.}

\item{dv}{AbstractVariable whose outcome we are interested in.}
}
\description{
This function disambiguates the Statistical Model.
}
