% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{Query a conceptual model for a statistical model}
\usage{
query(conceptualModel, iv, dv, data)
}
\arguments{
\item{conceptualModel}{ConceptualModel we are assuming in order to assess the effects of @param ivs on @param dv.}

\item{iv}{AbstractVariable. Which variable whose effect on @param dv we want to evaluate.}

\item{dv}{AbstractVariable. Variable whose outcome we want to assess.}
}
\description{
Method for querying a conceptual model for a statistical model
Fits and shows results of executing a statistical model + visualizing it from the conceptual model.
Returns an R script for running the statistical models + visualization
}
