% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\name{RandomIntercept}
\alias{RandomIntercept}
\title{RandomIntercept class.}
\usage{
RandomIntercept(group)
}
\description{
Class for declaring random intercepts
}
\section{Slots}{

\describe{
\item{\code{group}}{Unit whose observations we want to pool.}
}}

