% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\name{RandomSlope}
\alias{RandomSlope}
\title{RandomSlope class.}
\usage{
RandomSlope(variable, group)
}
\description{
Class for declaring random slopes
}
\section{Slots}{

\describe{
\item{\code{variable}}{Measure whose observations we want to calculate a slope for.}

\item{\code{group}}{Unit whose observations we want to pool.}
}}

