% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\docType{class}
\name{Relates-class}
\alias{Relates-class}
\alias{Relates}
\title{Relates class}
\description{
Class for Relates relationships.
Not called directly.
}
\section{Slots}{

\describe{
\item{\code{lhs}}{AbstractVariable. A variable.}

\item{\code{rhs}}{AbstractVariable. A variable.}
}}

