% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relates.R
\name{relates}
\alias{relates}
\title{Specify a relates (ambiguous) relationship}
\usage{
relates(lhs, rhs, when, then)
}
\arguments{
\item{lhs}{AbstractVariable. A variable.}

\item{rhs}{AbstractVariable. A variable.}
}
\description{
Method for constructing an ambiguous "relates" relationship between two variables.
Returns the Relates relationship constructed.
}
