% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\docType{class}
\name{SetUp-class}
\alias{SetUp-class}
\alias{SetUp}
\title{SetUp class}
\description{
Class for SetUp variables
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of SetUp, corresponds to column name if assigning data}

\item{\code{order}}{Optional. Order of categories if the SetUp variable represents an ordinal value (e.g., week of the month)}

\item{\code{cardinality}}{Optional. Cardinality of SetUp variable if itrepresents a nominal or ordinal value (e.g., trial identifier)}
}}

\examples{
SetUp()
}
