% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\name{Time}
\alias{Time}
\title{Time class}
\usage{
Time(name, order = list(), cardinality = as.integer(0))
}
\description{
Class for Time variables
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of Time, corresponds to column name if assigning data}

\item{\code{order}}{Optional. Order of categories if the Time variable represents an ordinal value (e.g., week of the month)}

\item{\code{cardinality}}{Optional. Cardinality of Time variable if itrepresents a nominal or ordinal value (e.g., trial identifier)}
}}

