% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processQuery.R
\name{updateConceptualModel}
\alias{updateConceptualModel}
\title{Updates Conceptual Model}
\usage{
updateConceptualModel(conceptualModel, values)
}
\arguments{
\item{conceptualModel}{ConceptualModel to update.}

\item{values}{ReactiveValues from disambiguating DV type.}
}
\description{
Updates Conceptual model based on analyst input during disambiguation
}
