% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processQuery.R
\name{updateDV}
\alias{updateDV}
\title{Updates DV type}
\usage{
updateDV(dv, values)
}
\arguments{
\item{dv}{AbstractVariable to cast.}

\item{values}{ReactiveValues from disambiguating DV type.}
}
\description{
Casts DV to be Continuous, Counts, or Categories, which is necessary to infer family and link functions
}
