% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateGraph.R
\name{updateGraph}
\alias{updateGraph}
\title{Update graph of all relationships in a Conceptual Model.}
\usage{
updateGraph(conceptualModel)
}
\arguments{
\item{conceptualModel}{ConceptualModel. Conceptual Model for which to create a causal graph.}
}
\value{
Dagitty DAG representing causal graph.
}
\description{
Constructs and returns a graph representing causal relationships.
}
