% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\docType{class}
\name{Categories-class}
\alias{Categories-class}
\alias{Categories}
\title{Categories class}
\description{
Wrapper class for Categorical measures.
Not called directly. All measures are declared through Units as either Numeric, Ordinal, or Nominal types.
}
\section{Slots}{

\describe{
\item{\code{measure}}{Measure to wrap.}

\item{\code{numberOfCategories}}{integer. Number of categories the measure contains.}
}}

