% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causes.R
\name{causes}
\alias{causes}
\title{Specify a causes relationship}
\usage{
causes(cause, effect, when, then)
}
\arguments{
\item{cause}{AbstractVariable. The cause.}

\item{effect}{AbstractVariable. The effect.}
}
\description{
Method for constructing a causes relationship between two variables.
Returns the Causes relationship constructed.
}
