% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\docType{class}
\name{Continuous-class}
\alias{Continuous-class}
\alias{Continuous}
\title{Continuous class}
\description{
Wrapper class for Continuous measures.
Not called directly. All measures are declared through Units as either Numeric, Ordinal, or Nominal types.
}
\section{Slots}{

\describe{
\item{\code{measure}}{Measure to wrap.}

\item{\code{skew}}{character. Description of skew, options are "positive" or NULL.}
}}

