% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\docType{class}
\name{Counts-class}
\alias{Counts-class}
\alias{Counts}
\title{Counts class}
\description{
Class for Counts
Inherits from Measure.
}
\section{Slots}{

\describe{
\item{\code{name}}{character. Name of measure, corresponds to column name in data.}
}}

