% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesize.R
\name{hypothesize}
\alias{hypothesize}
\title{Specify a Hypothesize statement}
\usage{
hypothesize(conceptualModel, relationship)
}
\arguments{
\item{conceptualModel}{ConceptualModel to which we should add the Hypothesized relationship.}

\item{relationship}{Relationship to assume.}
}
\description{
Method for constructing a hypothesis about a relationship.
Returns the Hypothesis created.
}
