% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferRandomEffects.R
\name{inferRandomEffects}
\alias{inferRandomEffects}
\title{Infers random effects.}
\usage{
inferRandomEffects(confounders, interactions, conceptualModel, iv, dv)
}
\arguments{
\item{confounders}{List of confounders to include in a statistical model.}

\item{interactions}{List of interaction terms to include in a statistical model.}

\item{conceptualModel}{ConceptualModel expressing conceptual relationships.}

\item{iv}{AbstractVariable whose influence on @param dv we are interested in estimating.}

\item{dv}{AbstractVariable whose outcome we are interested in estimating in a statsitical model.}
}
\value{
List of random effects to include.
}
\description{
This function infers random effects from data measurement relationships.
}
