% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\name{Unit}
\alias{Unit}
\title{Unit class}
\usage{
Unit(name, cardinality = as.integer(0), nestsWithin = NULL)
}
\description{
Class for Unit variables
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of Unit, corresponds to column name if assigning data}

\item{\code{integer}}{Integer for cardinality, optional. Only required if no data is assigned}

\item{\code{nestsWithin}}{Unit if this unit is nested within another (e.g., hierarchical relationship). Optional.}
}}

