% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whenThen.R
\name{whenThen}
\alias{whenThen}
\title{Specify a (hyper-)specific relationship about two or more relationships.}
\usage{
whenThen(when, then)
}
\arguments{
\item{when}{Compares or a list of Compares.}

\item{then}{Compares.}
}
\description{
Method for constructing a very specific relationship between two variables.
Returns the Causes relationship constructed if @param when is a single Compares/involves one variable.
Returns the Moderates relationship constructed if @param when is a list of Compares/involves multiple variables.
}
